requirements.txt
# Deep Learning and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Science and Machine Learning
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.1.0
scipy>=1.9.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.3.0
albumentations>=1.3.0

# Training and Logging
tensorboard>=2.10.0
wandb>=0.13.0
PyYAML>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
joblib>=1.0.0
tqdm>=4.64.0
pathlib2>=2.3.0
dataclasses>=0.8.0; python_version<"3.7"

# Jupyter
jupyter>=1.0.0
ipykernel>=6.15.0
nbconvert>=7.0.0

# Distributed Training
mpi4py>=3.1.0

# API and Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
aiohttp>=3.8.0
aiofiles>=0.8.0
python-multipart>=0.0.5

# System Monitoring
psutil>=5.8.0

# Model Export and Conversion
coremltools>=6.0.0
onnx==1.14.1
onnxruntime>=1.12.0
tensorflow>=2.8.0
onnx2tf>=1.16.31
ai-edge-torch>=0.1.0

# Windows Desktop Application
PyQt5>=5.15.0
pyinstaller>=5.0.0

# Medical Imaging (optional)
nibabel>=4.0.0
pydicom>=2.3.0